Create LocationCalculatorVääksy
The program utilizes the locations and types of apartments, along with road and walkway maps, as inputs. Additionally, it considers the current locations of services. It then approximates fuel usage for service locations in the city with adjusting density as citizens access nearby services. The output is a heatmap of the road network. I also collected data of Vääksy and ran the code to explore the outdoor gym locations in Vääksy.